% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_meta.R
\name{social_meta}
\alias{social_meta}
\title{Inject social metadata into Shiny UI}
\usage{
social_meta(meta)
}
\arguments{
\item{meta}{Either a path to a YAML file or a named list}
}
\description{
Inject social metadata into Shiny UI
}
